Rollback changes related to RandomizerFunnel
